A bit more refactorization
