+++ src/app/ui/page/about/about.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { version } from '../../../../package.json';
+import { version } from '../../../../../package.json';
 
 @Component({
   selector: 'app-about',
